FAD-5455 Restructure files, tweak row format
