+++ src/pages/credentials/components/KeyColumn.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const KeyColumn = ({ shortKey }) =>
+  <code>
+    {shortKey} ••••••••••••••••••••••••••••••••••••
+  </code>;
+
+export default KeyColumn;
