+++ src/pages/credentials/components/Permissions.module.scss
@@ -1,0 +1,3 @@
+.Permissions {
+   word-break: normal;
+}
